+++ src/utils/observableFromStore.js
@@ -1,6 +1,6 @@
 import {Observable} from 'rx-lite';
 
-export default function storeStream (store) {
+export default function observableFromStore (store) {
   return Observable.create(observer =>
     store.subscribe(() => observer.onNext(store.getState()))
   );
